Refactor API structure and add docker tooling
- refactor api structure
- add Dockerfile and docker-compose
