+++ .env.example
@@ -2,6 +2,7 @@ APP_ENV=local
 HOST=0.0.0.0
 PORT=8000
 CORS_ORIGINS=http://localhost:5173
+API_KEY=123456
 DB_HOST=
 DB_PORT=
 DB_NAME=
@@ -13,12 +14,12 @@ DB_OPTIONS=
 DATABASE_URL=sqlite:///instance/app.sqlite
 
 # PostgreSQL Configuration (for future chatbot features)
-# DATABASE_URL=postgresql://user:password@localhost:5432/chatbot_db
-# POSTGRES_HOST=localhost
-# POSTGRES_PORT=5432
-# POSTGRES_DB=chatbot_db
-# POSTGRES_USER=postgres
-# POSTGRES_PASSWORD=your_password
+DATABASE_URL=postgresql://admin:b@localhost:5432/ccc
+POSTGRES_HOST=localhost
+POSTGRES_PORT=5432
+POSTGRES_DB=ccc
+POSTGRES_USER=admin
+POSTGRES_PASSWORD=b
 
 # Webhook URLs
 WEBHOOK_UPLOAD_URL=https://iconic-host.lapage.vn/webhook/upload-file
