+++ api/routes/assignments.py
@@ -1,0 +1,113 @@
+﻿from __future__ import annotations
+
+import calendar
+import csv
+import io
+import json
+from datetime import date
+
+from flask import Blueprint, Response, jsonify, request, stream_with_context
+
+from api.export_month_csv import export_month_csv as export_month_csv_impl
+from models import Assignment, SessionLocal
+
+bp = Blueprint("assignments", __name__)
+
+
+@bp.get("/api/assignments")
+def list_assignments():
+    """Return assignments for the requested month (defaults to current)."""
+    today = date.today()
+    year = request.args.get("year", type=int) or today.year
+    month = request.args.get("month", type=int) or today.month
+
+    if month < 1 or month > 12:
+        return {"error": "month must be 1..12"}, 400
+
+    last_day = calendar.monthrange(year, month)[1]
+    start = date(year, month, 1)
+    end = date(year, month, last_day)
+
+    with SessionLocal() as session:
+        rows = (
+            session.query(Assignment)
+            .filter(Assignment.day.between(start, end))
+            .all()
+        )
+        return jsonify(
+            [
+                {
+                    "day": row.day.isoformat(),
+                    "shift_code": row.shift_code,
+                    "staff_id": row.staff_id,
+                    "position": row.position,
+                }
+                for row in rows
+            ]
+        )
+
+
+@bp.get("/api/export_audit")
+def export_audit():
+    """Export assignments with metadata for auditing in CSV format."""
+    today = date.today()
+    year = request.args.get("year", type=int) or today.year
+    month = request.args.get("month", type=int) or today.month
+
+    if month < 1 or month > 12:
+        return {"error": "month must be 1..12"}, 400
+
+    last_day = calendar.monthrange(year, month)[1]
+    start = date(year, month, 1)
+    end = date(year, month, last_day)
+
+    with SessionLocal() as session:
+        rows = (
+            session.query(Assignment)
+            .filter(Assignment.day.between(start, end))
+            .all()
+        )
+
+    def generate():
+        buf = io.StringIO()
+        writer = csv.writer(buf)
+        writer.writerow(["day", "shift_code", "staff_id", "meta"])
+        yield buf.getvalue()
+        buf.seek(0)
+        buf.truncate(0)
+        for row in rows:
+            meta = {
+                "info": f"{row.shift_code},{row.staff_id}",
+                "extra": "value",
+            }
+            writer.writerow(
+                [
+                    row.day.isoformat(),
+                    row.shift_code,
+                    row.staff_id,
+                    json.dumps(meta, ensure_ascii=False),
+                ]
+            )
+            yield buf.getvalue()
+            buf.seek(0)
+            buf.truncate(0)
+
+    headers = {
+        "Content-Disposition": f"attachment; filename=audit-{year:04d}-{month:02d}.csv"
+    }
+    return Response(
+        stream_with_context(generate()),
+        content_type="text/csv; charset=utf-8",
+        headers=headers,
+    )
+
+
+@bp.get("/api/export/month.csv")
+def export_month_csv_endpoint():
+    """Export the monthly schedule as CSV."""
+    today = date.today()
+    year = request.args.get("year", type=int) or today.year
+    month = request.args.get("month", type=int) or today.month
+    if month < 1 or month > 12:
+        return {"error": "month must be 1..12"}, 400
+    return export_month_csv_impl(year, month)
