+++ app.py
@@ -1,59 +1,27 @@
-# backend/app.py
-import calendar
-import csv
-import io
-import json
+﻿# backend/app.py
 import os
 import pathlib
-import urllib.error
-import urllib.request
-from collections import defaultdict
-from datetime import date, timedelta
 
-from flask import Flask, Response, jsonify, request, send_from_directory, stream_with_context
+from flask import Flask
 from flask_cors import CORS
-from sqlalchemy import func, select, text
-from sqlalchemy.orm import selectinload
-
 from dotenv import load_dotenv
+
+from api.routes import register_routes
+from middleware.api_key import register_api_key_middleware
+from models import init_db
 
 load_dotenv()
 
-from api.export_month_csv import export_month_csv as _export_month_csv
-from api.metrics import load_department_comparison, load_staff_workload
-from models import (
-    Assignment,
-    Department,
-    FixedAssignment,
-    Holiday,
-    MonthConfig,
-    OffDay,
-    ShiftConfig,
-    ShiftPlanDefaults,
-    SessionLocal,
-    Staff,
-    WeekendPolicy,
-    engine,
-    init_db,
-)
-from rules import SHIFT_DEFS, get_profile
-from scheduler import schedule_month as generate_schedule
-from scheduler.estimator import estimate_month
-from scheduler.repo import load_holidays
-from scheduler.utils import day_kind, month_last_day, ymd
-from scheduler.validate import validate_month
-
-# Trỏ tới thư mục build của frontend (Vite) nếu đã build
+# Configure static asset serving (frontend build directory)
 BASE_DIR = pathlib.Path(__file__).parent
 DEFAULT_FRONTEND_DIST = BASE_DIR.parent / "frontend" / "dist"
 frontend_override = os.getenv("FRONTEND_DIST")
-frontend_path = pathlib.Path(frontend_override).expanduser() if frontend_override else DEFAULT_FRONTEND_DIST
+frontend_path = (
+    pathlib.Path(frontend_override).expanduser()
+    if frontend_override
+    else DEFAULT_FRONTEND_DIST
+)
 frontend_static = str(frontend_path.resolve()) if frontend_path.exists() else None
-
-db_path = pathlib.Path(engine.url.database).expanduser() if engine.url.database else None
-if db_path and not db_path.is_absolute():
-    db_path = (pathlib.Path.cwd() / db_path).resolve()
-DB_FILE = db_path
 
 app = Flask(
     __name__,
@@ -67,2294 +35,16 @@ CORS(app, origins=cors_origins or ["*"], supports_credentials=True)
 APP_HOST = os.getenv("HOST", "0.0.0.0")
 APP_PORT = int(os.getenv("PORT", "8000"))
 APP_DEBUG = os.getenv("APP_ENV", "local").lower() != "production"
+API_KEY = os.getenv("API_KEY")
 
-# Khởi tạo DB (tạo bảng nếu chưa có)
+# Initialise database schema
 init_db()
 
-
-class HolidayImportError(RuntimeError):
-    """Raised when an external holiday provider cannot be fetched."""
-
-
-def _serialize_holiday(row: Holiday) -> dict:
-    """Convert a Holiday row to JSON-serialisable dict."""
-
-    return {
-        "id": row.id,
-        "day": row.day.isoformat(),
-        "name": row.name,
-        "kind": row.kind,
-        "official": bool(row.official),
-        "source": row.source,
-    }
-
-
-def _fetch_nager_holidays(year: int, country_code: str = "VN") -> list[dict]:
-    """Fetch holidays from Nager.Date API for a year."""
-
-    url = f"https://date.nager.at/api/v3/PublicHolidays/{year}/{country_code}".strip()
-    try:
-        with urllib.request.urlopen(url, timeout=15) as response:
-            status = getattr(response, "status", None) or response.getcode()
-            if status >= 400:
-                raise HolidayImportError(f"Provider returned HTTP {status}")
-            payload = response.read()
-    except urllib.error.URLError as exc:  # pragma: no cover - network failure branch
-        raise HolidayImportError(str(exc)) from exc
-
-    try:
-        data = json.loads(payload)
-    except json.JSONDecodeError as exc:  # pragma: no cover - provider bug branch
-        raise HolidayImportError("Invalid JSON payload") from exc
-
-    if not isinstance(data, list):
-        raise HolidayImportError("Unexpected response schema")
-
-    out: list[dict] = []
-    for item in data:
-        if not isinstance(item, dict):
-            continue
-        day = item.get("date")
-        try:
-            day_value = date.fromisoformat(day)
-        except Exception:
-            continue
-        name = item.get("localName") or item.get("name") or "(unnamed)"
-        types = item.get("types") or []
-        kind = ",".join(t for t in types if isinstance(t, str)) or None
-        official = "Public" in types or bool(item.get("global"))
-        out.append(
-            {
-                "day": day_value,
-                "name": name,
-                "kind": kind,
-                "official": official,
-                "source": "nager",
-            }
-        )
-    return out
-
-
-def _parse_year_month(year_raw, month_raw) -> tuple[int, int]:
-    """Validate and convert year/month inputs."""
-
-    if year_raw is None or month_raw is None:
-        raise ValueError("year and month required")
-    try:
-        year = int(year_raw)
-        month = int(month_raw)
-    except (TypeError, ValueError) as exc:  # pragma: no cover - defensive guard
-        raise ValueError("year and month must be integers") from exc
-    if month < 1 or month > 12:
-        raise ValueError("month must be between 1 and 12")
-    if year < 1900 or year > 2100:
-        raise ValueError("year must be between 1900 and 2100")
-    return year, month
-
-
-def _get_cost_per_hour() -> float:
-    raw = os.getenv("LABOR_COST_PER_HOUR", "0").strip()
-    try:
-        return float(raw)
-    except (TypeError, ValueError):
-        return 0.0
-
-
-def _build_csv_response(filename: str, header: list[str], rows: list[list[object]]):
-    def generate():
-        buf = io.StringIO()
-        writer = csv.writer(buf)
-        writer.writerow(header)
-        yield buf.getvalue()
-        buf.seek(0)
-        buf.truncate(0)
-        for row in rows:
-            writer.writerow(row)
-            yield buf.getvalue()
-            buf.seek(0)
-            buf.truncate(0)
-
-    response = Response(
-        stream_with_context(generate()),
-        content_type="text/csv; charset=utf-8",
-    )
-    response.headers["Content-Disposition"] = f"attachment; filename={filename}"
-    return response
-
-
-def _coerce_extra_dates(
-    values: object, year: int, month: int, field_name: str
-) -> set[date] | None:
-    """Validate optional list of YYYY-MM-DD strings limited to the target month."""
-
-    if values is None:
-        return None
-    if not isinstance(values, list):
-        raise ValueError(f"{field_name} must be a list of dates")
-
-    parsed: set[date] = set()
-    for item in values:
-        if not isinstance(item, str):
-            raise ValueError(f"{field_name} must contain YYYY-MM-DD strings")
-        try:
-            dt = date.fromisoformat(item)
-        except ValueError as exc:
-            raise ValueError(f"{field_name} has invalid date '{item}'") from exc
-        if dt.year != year or dt.month != month:
-            raise ValueError(
-                f"{field_name} must stay within {year}-{month:02d}"
-            )
-        parsed.add(dt)
-    return parsed
-
-
-def _build_month_config_payload(session, year: int, month: int) -> dict:
-    """Construct payload for month config, including derived working days."""
-
-    config = session.execute(
-        select(MonthConfig).where(
-            MonthConfig.year == year, MonthConfig.month == month
-        )
-    ).scalar_one_or_none()
-
-    weekend_policy = (
-        config.weekend_policy if config else WeekendPolicy.SAT_OFF
-    )
-    stored_workdays = config.extra_workdays if config else []
-    stored_offdays = config.extra_offdays if config else []
-    working_days_override = (
-        config.working_days_override if config else None
-    )
-
-    last_day = month_last_day(year, month)
-    start = date(year, month, 1)
-    end = date(year, month, last_day)
-
-    official_holidays = {
-        row.day
-        for row in session.scalars(
-            select(Holiday).where(
-                Holiday.day >= start,
-                Holiday.day <= end,
-                Holiday.official.is_(True),
-            )
-        )
-    }
-
-    auto_working_days = sum(
-        1
-        for day in range(1, last_day + 1)
-        if (dt := date(year, month, day)).weekday() < 5
-        and dt not in official_holidays
-    )
-
-    day_values: dict[date, float] = {}
-    for day in range(1, last_day + 1):
-        current = date(year, month, day)
-        weekday = current.weekday()
-        if weekday < 5:
-            base = 1.0
-        elif weekday == 5:
-            if weekend_policy == WeekendPolicy.SAT_WORK:
-                base = 1.0
-            elif weekend_policy == WeekendPolicy.SAT_WORK_AM:
-                base = 0.5
-            else:
-                base = 0.0
-        else:
-            base = 0.0
-        if current in official_holidays:
-            base = 0.0
-        day_values[current] = base
-
-    workday_dates = set()
-    for raw in stored_workdays:
-        try:
-            dt = date.fromisoformat(raw)
-        except ValueError:
-            continue
-        if dt.year == year and dt.month == month:
-            workday_dates.add(dt)
-    offday_dates = set()
-    for raw in stored_offdays:
-        try:
-            dt = date.fromisoformat(raw)
-        except ValueError:
-            continue
-        if dt.year == year and dt.month == month:
-            offday_dates.add(dt)
-
-    for extra in workday_dates:
-        if start <= extra <= end:
-            day_values[extra] = max(day_values.get(extra, 0.0), 1.0)
-
-    for extra in offday_dates:
-        if start <= extra <= end:
-            day_values[extra] = 0.0
-
-    policy_working_days = sum(day_values.values())
-    effective_working_days = (
-        working_days_override
-        if working_days_override is not None
-        else policy_working_days
-    )
-
-    payload = {
-        "year": year,
-        "month": month,
-        "auto_working_days": auto_working_days,
-        "policy_working_days": policy_working_days,
-        "effective_working_days": effective_working_days,
-        "config": {
-            "weekend_policy": weekend_policy.value,
-            "extra_workdays": sorted(dt.isoformat() for dt in workday_dates),
-            "extra_offdays": sorted(dt.isoformat() for dt in offday_dates),
-            "working_days_override": working_days_override,
-        },
-    }
-    return payload
-
-
-def _serialize_shift_defaults(
-    defaults: ShiftPlanDefaults | None, year: int, month: int
-) -> dict:
-    """Return API payload describing stored shift plan defaults for a month."""
-
-    return {
-        "year": year,
-        "month": month,
-        "day_shifts": defaults.day_shifts if defaults else 0,
-        "night_shifts": defaults.night_shifts if defaults else 0,
-        "leader_shifts": defaults.leader_shifts if defaults else 0,
-        "pgd_shifts": defaults.pgd_shifts if defaults else 0,
-    }
-
-
-# ---------- Root / SPA ----------
-@app.get("/")
-def root():
-    """Nếu đã build frontend thì trả index.html, còn không trả JSON giới thiệu API."""
-    if app.static_folder and os.path.isfile(os.path.join(app.static_folder, "index.html")):
-        return send_from_directory(app.static_folder, "index.html")
-    return jsonify(
-        service="customer-care-center API",
-        ok=True,
-        docs=[
-            "/api/ping",
-            "/api/shifts",
-            "/api/staff",
-            "/api/assignments?year=YYYY&month=MM",
-            "/api/schedule/generate (POST)",
-            "/api/admin/reset?mode=soft (POST)",
-        ],
-    )
-
-
-@app.route("/<path:path>")
-def spa_assets(path: str):
-    """Phục vụ static assets và fallback cho route SPA."""
-    if app.static_folder:
-        p = os.path.join(app.static_folder, path)
-        if os.path.isfile(p):
-            return send_from_directory(app.static_folder, path)
-        index_html = os.path.join(app.static_folder, "index.html")
-        if os.path.isfile(index_html):
-            return send_from_directory(app.static_folder, "index.html")
-    return {"error": "Not Found"}, 404
-
-
-# ---------- API ----------
-@app.get("/api/ping")
-def ping():
-    return {"ok": True}
-
-
-# ===========================================
-# Department Management APIs
-# ===========================================
-
-@app.get("/api/departments")
-def list_departments():
-    """Get all departments."""
-
-    active_raw = request.args.get("active")
-    active_filter: bool | None = None
-    if active_raw is not None:
-        raw = active_raw.strip().lower()
-        if raw in {"1", "true", "yes"}:
-            active_filter = True
-        elif raw in {"0", "false", "no"}:
-            active_filter = False
-        else:
-            return jsonify({"error": "active must be truthy (1/true) or falsy (0/false)"}), 400
-
-    with SessionLocal() as s:
-        query = select(Department).order_by(Department.name)
-        if active_filter is not None:
-            query = query.where(Department.is_active.is_(active_filter))
-        rows = s.execute(query).scalars().all()
-
-        if active_filter is not None:
-            payload = [
-                {
-                    "id": r.id,
-                    "name": r.name,
-                    "code": r.code,
-                    "color": r.color,
-                    "icon": r.icon,
-                    "settings": r.settings,
-                }
-                for r in rows
-            ]
-        else:
-            payload = [
-                {
-                    "id": r.id,
-                    "name": r.name,
-                    "code": r.code,
-                    "color": r.color,
-                    "icon": r.icon,
-                    "description": r.description,
-                    "is_active": r.is_active,
-                    "settings": r.settings,
-                    "staff_count": len(r.staff) if r.staff else 0,
-                    "shift_count": len(r.shifts) if r.shifts else 0,
-                }
-                for r in rows
-            ]
-
-        return jsonify(payload)
-
-
-@app.post("/api/departments")
-def create_department():
-    """Create a new department."""
-    data = request.get_json() or {}
-    name = data.get("name", "").strip()
-    code = data.get("code", "").strip().upper()
-
-    if not name:
-        return jsonify({"error": "Department name is required"}), 400
-    if not code:
-        return jsonify({"error": "Department code is required"}), 400
-
-    with SessionLocal() as s:
-        # Check duplicates
-        existing = s.execute(
-            select(Department).where(
-                (Department.name == name) | (Department.code == code)
-            )
-        ).scalar_one_or_none()
-
-        if existing:
-            return jsonify({"error": "Department name or code already exists"}), 400
-
-        dept = Department(
-            name=name,
-            code=code,
-            color=data.get("color", "#3b82f6"),
-            icon=data.get("icon", "Building2"),
-            description=data.get("description"),
-            is_active=data.get("is_active", True),
-            settings=data.get("settings", {
-                "working_hours": {"start": "08:00", "end": "17:00"},
-                "weekend_policy": "sat_off",
-                "max_hours_per_month": 208,
-                "min_staff_per_shift": 2,
-            }),
-        )
-        s.add(dept)
-        s.commit()
-        s.refresh(dept)
-
-        return jsonify({
-            "id": dept.id,
-            "name": dept.name,
-            "code": dept.code,
-            "color": dept.color,
-            "icon": dept.icon,
-            "description": dept.description,
-            "is_active": dept.is_active,
-            "settings": dept.settings,
-        }), 201
-
-
-@app.put("/api/departments/<int:dept_id>")
-def update_department(dept_id: int):
-    """Update a department."""
-    data = request.get_json() or {}
-
-    with SessionLocal() as s:
-        dept = s.get(Department, dept_id)
-        if not dept:
-            return jsonify({"error": "Department not found"}), 404
-
-        # Update fields
-        if "name" in data:
-            name = data["name"].strip()
-            if not name:
-                return jsonify({"error": "Department name cannot be empty"}), 400
-            dept.name = name
-
-        if "code" in data:
-            code = data["code"].strip().upper()
-            if not code:
-                return jsonify({"error": "Department code cannot be empty"}), 400
-            dept.code = code
-
-        if "color" in data:
-            dept.color = data["color"]
-        if "icon" in data:
-            dept.icon = data["icon"]
-        if "description" in data:
-            dept.description = data["description"]
-        if "is_active" in data:
-            dept.is_active = data["is_active"]
-        if "settings" in data:
-            dept.settings = data["settings"]
-
-        s.commit()
-        s.refresh(dept)
-
-        return jsonify({
-            "id": dept.id,
-            "name": dept.name,
-            "code": dept.code,
-            "color": dept.color,
-            "icon": dept.icon,
-            "description": dept.description,
-            "is_active": dept.is_active,
-            "settings": dept.settings,
-        })
-
-
-@app.delete("/api/departments/<int:dept_id>")
-def delete_department(dept_id: int):
-    """Delete a department."""
-    with SessionLocal() as s:
-        dept = s.get(Department, dept_id)
-        if not dept:
-            return jsonify({"error": "Department not found"}), 404
-
-        # Check if department has staff
-        if dept.staff and len(dept.staff) > 0:
-            return jsonify({
-                "error": f"Cannot delete department with {len(dept.staff)} staff members. Please reassign or remove staff first."
-            }), 400
-
-        s.delete(dept)
-        s.commit()
-        return jsonify({"ok": True})
-
-
-# ===========================================
-# Shift Config APIs
-# ===========================================
-
-@app.get("/api/shift-configs")
-def list_shift_configs():
-    """Get all shift configurations, optionally filtered by department."""
-    dept_id = request.args.get("department_id", type=int)
-
-    with SessionLocal() as s:
-        query = select(ShiftConfig).order_by(
-            ShiftConfig.department_id,
-            ShiftConfig.display_order,
-            ShiftConfig.name
-        )
-
-        if dept_id:
-            query = query.where(ShiftConfig.department_id == dept_id)
-
-        rows = s.execute(query).scalars().all()
-        return jsonify(
-            [
-                {
-                    "id": r.id,
-                    "department_id": r.department_id,
-                    "department_name": r.department.name if r.department else None,
-                    "name": r.name,
-                    "code": r.code,
-                    "start_time": r.start_time,
-                    "end_time": r.end_time,
-                    "color": r.color,
-                    "icon": r.icon,
-                    "is_active": r.is_active,
-                    "display_order": r.display_order,
-                    "rules": r.rules,
-                }
-                for r in rows
-            ]
-        )
-
-
-@app.post("/api/shift-configs")
-def create_shift_config():
-    """Create a new shift configuration."""
-    data = request.get_json() or {}
-    dept_id = data.get("department_id")
-    name = data.get("name", "").strip()
-    code = data.get("code", "").strip().upper()
-
-    if not dept_id:
-        return jsonify({"error": "department_id is required"}), 400
-    if not name:
-        return jsonify({"error": "Shift name is required"}), 400
-    if not code:
-        return jsonify({"error": "Shift code is required"}), 400
-
-    with SessionLocal() as s:
-        # Check if department exists
-        dept = s.get(Department, dept_id)
-        if not dept:
-            return jsonify({"error": "Department not found"}), 404
-
-        # Check for duplicate code in same department
-        existing = s.execute(
-            select(ShiftConfig).where(
-                (ShiftConfig.department_id == dept_id) &
-                (ShiftConfig.code == code)
-            )
-        ).scalar_one_or_none()
-
-        if existing:
-            return jsonify({"error": f"Shift code '{code}' already exists in this department"}), 400
-
-        shift = ShiftConfig(
-            department_id=dept_id,
-            name=name,
-            code=code,
-            start_time=data.get("start_time", "08:00"),
-            end_time=data.get("end_time", "17:00"),
-            color=data.get("color", "#60a5fa"),
-            icon=data.get("icon", "Sun"),
-            is_active=data.get("is_active", True),
-            display_order=data.get("display_order", 0),
-            rules=data.get("rules", {}),
-        )
-        s.add(shift)
-        s.commit()
-        s.refresh(shift)
-
-        return jsonify({
-            "id": shift.id,
-            "department_id": shift.department_id,
-            "name": shift.name,
-            "code": shift.code,
-            "start_time": shift.start_time,
-            "end_time": shift.end_time,
-            "color": shift.color,
-            "icon": shift.icon,
-            "is_active": shift.is_active,
-            "display_order": shift.display_order,
-            "rules": shift.rules,
-        }), 201
-
-
-@app.put("/api/shift-configs/<int:shift_id>")
-def update_shift_config(shift_id: int):
-    """Update a shift configuration."""
-    data = request.get_json() or {}
-
-    with SessionLocal() as s:
-        shift = s.get(ShiftConfig, shift_id)
-        if not shift:
-            return jsonify({"error": "Shift configuration not found"}), 404
-
-        # Update fields
-        if "name" in data:
-            shift.name = data["name"].strip()
-        if "code" in data:
-            shift.code = data["code"].strip().upper()
-        if "start_time" in data:
-            shift.start_time = data["start_time"]
-        if "end_time" in data:
-            shift.end_time = data["end_time"]
-        if "color" in data:
-            shift.color = data["color"]
-        if "icon" in data:
-            shift.icon = data["icon"]
-        if "is_active" in data:
-            shift.is_active = data["is_active"]
-        if "display_order" in data:
-            shift.display_order = data["display_order"]
-        if "rules" in data:
-            shift.rules = data["rules"]
-
-        s.commit()
-        s.refresh(shift)
-
-        return jsonify({
-            "id": shift.id,
-            "department_id": shift.department_id,
-            "name": shift.name,
-            "code": shift.code,
-            "start_time": shift.start_time,
-            "end_time": shift.end_time,
-            "color": shift.color,
-            "icon": shift.icon,
-            "is_active": shift.is_active,
-            "display_order": shift.display_order,
-            "rules": shift.rules,
-        })
-
-
-@app.delete("/api/shift-configs/<int:shift_id>")
-def delete_shift_config(shift_id: int):
-    """Delete a shift configuration."""
-    with SessionLocal() as s:
-        shift = s.get(ShiftConfig, shift_id)
-        if not shift:
-            return jsonify({"error": "Shift configuration not found"}), 404
-
-        s.delete(shift)
-        s.commit()
-        return jsonify({"ok": True})
-
-
-@app.get("/api/shifts")
-def shifts():
-    return jsonify(SHIFT_DEFS)
-
-
-@app.get("/api/staff")
-def list_staff():
-    """Get all staff, optionally filtered by department."""
-    dept_id = request.args.get("department_id", type=int)
-    role_filter_raw = request.args.get("role")
-    query_raw = request.args.get("q", "").strip()
-
-    with SessionLocal() as s:
-        query = select(Staff).options(selectinload(Staff.department))
-
-        if dept_id is not None:
-            query = query.where(Staff.department_id == dept_id)
-
-        if role_filter_raw:
-            roles = {item.strip().upper() for item in role_filter_raw.split(",") if item.strip()}
-            if roles:
-                query = query.where(func.upper(Staff.role).in_(sorted(roles)))
-
-        if query_raw:
-            pattern = f"%{query_raw.lower()}%"
-            query = query.where(func.lower(Staff.full_name).like(pattern))
-
-        query = query.order_by(func.lower(Staff.full_name))
-
-        rows = s.execute(query).scalars().all()
-        return jsonify(
-            [
-                {
-                    "id": r.id,
-                    "full_name": r.full_name,
-                    "role": r.role,
-                    "can_night": r.can_night,
-                    "base_quota": r.base_quota,
-                    "notes": r.notes,
-                    "department_id": r.department_id,
-                    "department_name": r.department.name if r.department else None,
-                }
-                for r in rows
-            ]
-        )
-
-
-def _month_range(year: int, month: int) -> tuple[date, date]:
-    last_day = month_last_day(year, month)
-    start = date(year, month, 1)
-    end = date(year, month, last_day)
-    return start, end
-
-
-@app.get("/api/schedule")
-def get_schedule():
-    """Return assignments for a month with optional department filtering."""
-
-    try:
-        year, month = _parse_year_month(request.args.get("year"), request.args.get("month"))
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-
-    dept_id = request.args.get("department_id", type=int)
-    start, end = _month_range(year, month)
-
-    with SessionLocal() as s:
-        if dept_id is not None:
-            dept = s.get(Department, dept_id)
-            if not dept or not dept.is_active:
-                return {"error": "Department not found or inactive"}, 404
-
-        query = (
-            select(
-                Assignment.day,
-                Assignment.shift_code,
-                Assignment.position,
-                Assignment.staff_id,
-                Staff.full_name,
-                Staff.role,
-                Staff.department_id,
-                Department.name,
-            )
-            .join(Staff, Assignment.staff_id == Staff.id, isouter=True)
-            .join(Department, Staff.department_id == Department.id, isouter=True)
-            .where(Assignment.day.between(start, end))
-            .order_by(Assignment.day, Assignment.shift_code, Assignment.id)
-        )
-
-        if dept_id is not None:
-            query = query.where(Staff.department_id == dept_id)
-
-        rows = s.execute(query).all()
-
-    items: list[dict] = []
-    counts: dict[str, object] = {
-        "total": 0,
-        "by_shift": defaultdict(int),
-        "leaders": {"day": 0, "night": 0},
-    }
-
-    for day, shift_code, position, staff_id, full_name, role, row_dept_id, dept_name in rows:
-        items.append(
-            {
-                "day": day.isoformat(),
-                "shift_code": shift_code,
-                "position": position,
-                "staff_id": staff_id,
-                "staff_name": full_name,
-                "role": role,
-                "department_id": row_dept_id,
-                "department_name": dept_name,
-            }
-        )
-        counts["total"] += 1
-        counts["by_shift"][shift_code] += 1
-        if shift_code == "K" and (position or "").upper() == "TD":
-            counts["leaders"]["day"] += 1
-        if shift_code == "Đ" and (position or "").upper() == "TD":
-            counts["leaders"]["night"] += 1
-
-    counts["by_shift"] = dict(counts["by_shift"])
-
-    return jsonify({"items": items, "counts": counts})
-
-
-@app.get("/api/schedule/overview")
-def schedule_overview():
-    """Summarise schedule coverage per department for a month."""
-
-    try:
-        year, month = _parse_year_month(request.args.get("year"), request.args.get("month"))
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-
-    start, end = _month_range(year, month)
-    days_in_month = month_last_day(year, month)
-
-    with SessionLocal() as s:
-        departments = (
-            s.execute(select(Department).order_by(Department.name)).scalars().all()
-        )
-        rows = s.execute(
-            select(
-                Assignment.day,
-                Assignment.shift_code,
-                Assignment.position,
-                Staff.department_id,
-                Department.name,
-                Department.is_active,
-            )
-            .join(Staff, Assignment.staff_id == Staff.id, isouter=True)
-            .join(Department, Staff.department_id == Department.id, isouter=True)
-            .where(Assignment.day.between(start, end))
-        ).all()
-
-    overview: dict[int, dict[str, object]] = {}
-    day_presence: dict[int, set[date]] = defaultdict(set)
-    day_shifts: dict[int, dict[str, set[date]]] = defaultdict(lambda: {"day": set(), "night": set()})
-    leader_flags: dict[int, dict[str, set[date]]] = defaultdict(lambda: {"day": set(), "night": set()})
-
-    for dept in departments:
-        if not dept.is_active:
-            continue
-        overview[dept.id] = {
-            "department_id": dept.id,
-            "name": dept.name,
-            "shifts": 0,
-            "missing_leaders": 0,
-            "coverage_rate": 0.0,
-        }
-
-    for day, shift_code, position, dept_id, dept_name, is_active in rows:
-        if dept_id is None or dept_id not in overview:
-            continue
-        info = overview[dept_id]
-        info["shifts"] += 1
-        day_presence[dept_id].add(day)
-        if shift_code == "K":
-            day_shifts[dept_id]["day"].add(day)
-            if (position or "").upper() == "TD":
-                leader_flags[dept_id]["day"].add(day)
-        elif shift_code == "Đ":
-            day_shifts[dept_id]["night"].add(day)
-            if (position or "").upper() == "TD":
-                leader_flags[dept_id]["night"].add(day)
-
-    for dept_id, info in overview.items():
-        covered_days = len(day_presence.get(dept_id, set()))
-        info["coverage_rate"] = round(
-            covered_days / days_in_month if days_in_month else 0.0,
-            4,
-        )
-
-        missing = 0
-        for bucket in ("day", "night"):
-            shift_days = day_shifts[dept_id][bucket]
-            leaders = leader_flags[dept_id][bucket]
-            for day in shift_days:
-                if day not in leaders:
-                    missing += 1
-        info["missing_leaders"] = missing
-
-    return jsonify(sorted(overview.values(), key=lambda item: item["name"]))
-
-
-@app.post("/api/staff")
-def add_staff():
-    data = request.get_json(force=True)
-    with SessionLocal() as s:
-        r = Staff(
-            full_name=data["full_name"],
-            role=data.get("role", "GDV"),
-            can_night=bool(data.get("can_night", True)),
-            base_quota=float(data.get("base_quota", 26.0)),
-            notes=data.get("notes"),
-            department_id=data.get("department_id"),
-        )
-        s.add(r)
-        s.commit()
-        s.refresh(r)
-        return {
-            "id": r.id,
-            "full_name": r.full_name,
-            "department_id": r.department_id,
-        }
-
-
-@app.put("/api/staff/<int:staff_id>")
-def update_staff(staff_id: int):
-    """Update an existing staff member."""
-    data = request.get_json(force=True)
-    with SessionLocal() as s:
-        r = s.get(Staff, staff_id)
-        if not r:
-            return {"error": "Staff not found"}, 404
-
-        # Update fields
-        if "full_name" in data:
-            r.full_name = data["full_name"]
-        if "role" in data:
-            r.role = data["role"]
-        if "can_night" in data:
-            r.can_night = bool(data["can_night"])
-        if "base_quota" in data:
-            r.base_quota = float(data["base_quota"])
-        if "notes" in data:
-            r.notes = data["notes"]
-        if "department_id" in data:
-            r.department_id = data["department_id"]
-
-        s.commit()
-        s.refresh(r)
-        return {
-            "id": r.id,
-            "full_name": r.full_name,
-            "role": r.role,
-            "can_night": r.can_night,
-            "base_quota": r.base_quota,
-            "notes": r.notes,
-            "department_id": r.department_id,
-            "department_name": r.department.name if r.department else None,
-        }
-
-
-@app.delete("/api/staff/<int:staff_id>")
-def del_staff(staff_id: int):
-    with SessionLocal() as s:
-        r = s.get(Staff, staff_id)
-        if not r:
-            return {"error": "Not found"}, 404
-        s.delete(r)
-        s.commit()
-        return {"ok": True}
-
-
-@app.get("/api/staff/<int:staff_id>/preferences")
-def get_staff_preferences(staff_id: int):
-    """Get scheduling preferences for a staff member."""
-    with SessionLocal() as s:
-        staff = s.get(Staff, staff_id)
-        if not staff:
-            return {"error": "Staff not found"}, 404
-
-        prefs = staff.preferences
-        if not prefs:
-            # Return default empty preferences
-            return {
-                "staff_id": staff_id,
-                "preferred_shifts": [],
-                "unavailable_days": [],
-                "max_consecutive_days": 6,
-                "preferred_days_off": [],
-                "notes": None,
-            }
-
-        return {
-            "staff_id": prefs.staff_id,
-            "preferred_shifts": prefs.preferred_shifts or [],
-            "unavailable_days": prefs.unavailable_days or [],
-            "max_consecutive_days": prefs.max_consecutive_days,
-            "preferred_days_off": prefs.preferred_days_off or [],
-            "notes": prefs.notes,
-        }
-
-
-@app.put("/api/staff/<int:staff_id>/preferences")
-def update_staff_preferences(staff_id: int):
-    """Upsert scheduling preferences for a staff member."""
-    from models import StaffPreferences
-    from datetime import datetime
-
-    data = request.get_json(force=True)
-
-    # Validation
-    preferred_shifts = data.get("preferred_shifts", [])
-    unavailable_days = data.get("unavailable_days", [])
-    max_consecutive_days = data.get("max_consecutive_days")
-    preferred_days_off = data.get("preferred_days_off", [])
-    notes = data.get("notes")
-
-    # Validate max_consecutive_days
-    if max_consecutive_days is not None:
-        try:
-            max_consecutive_days = int(max_consecutive_days)
-            if max_consecutive_days < 0:
-                return {"error": "max_consecutive_days must be >= 0"}, 400
-        except (ValueError, TypeError):
-            return {"error": "max_consecutive_days must be a number"}, 400
-
-    # Validate unavailable_days (ISO date format)
-    for day_str in unavailable_days:
-        try:
-            datetime.fromisoformat(day_str)
-        except (ValueError, TypeError):
-            return {"error": f"Invalid date format: {day_str}. Use YYYY-MM-DD"}, 400
-
-    # Validate preferred_days_off (0-6)
-    for day_num in preferred_days_off:
-        if not isinstance(day_num, int) or day_num < 0 or day_num > 6:
-            return {"error": f"Invalid day number: {day_num}. Must be 0-6 (Mon-Sun)"}, 400
-
-    # Validate preferred_shifts against department's shift configs
-    with SessionLocal() as s:
-        staff = s.get(Staff, staff_id)
-        if not staff:
-            return {"error": "Staff not found"}, 404
-
-        if staff.department_id and preferred_shifts:
-            from sqlalchemy import select
-            from models import ShiftConfig
-
-            valid_shift_codes = set(
-                s.execute(
-                    select(ShiftConfig.code)
-                    .where(ShiftConfig.department_id == staff.department_id)
-                    .where(ShiftConfig.is_active == True)
-                ).scalars().all()
-            )
-
-            for shift_code in preferred_shifts:
-                if shift_code not in valid_shift_codes:
-                    return {
-                        "error": f"Invalid shift code '{shift_code}' for department. Valid codes: {sorted(valid_shift_codes)}"
-                    }, 400
-
-        # Upsert preferences
-        prefs = staff.preferences
-        if not prefs:
-            prefs = StaffPreferences(staff_id=staff_id)
-            s.add(prefs)
-
-        prefs.preferred_shifts = preferred_shifts
-        prefs.unavailable_days = unavailable_days
-        prefs.max_consecutive_days = max_consecutive_days
-        prefs.preferred_days_off = preferred_days_off
-        prefs.notes = notes
-
-        s.commit()
-        s.refresh(prefs)
-
-        return {
-            "staff_id": prefs.staff_id,
-            "preferred_shifts": prefs.preferred_shifts or [],
-            "unavailable_days": prefs.unavailable_days or [],
-            "max_consecutive_days": prefs.max_consecutive_days,
-            "preferred_days_off": prefs.preferred_days_off or [],
-            "notes": prefs.notes,
-        }
-
-
-@app.get("/api/fixed")
-def list_fixed_assignments():
-    """List fixed assignments for a given month."""
-    today = date.today()
-    y = request.args.get("year", type=int) or today.year
-    m = request.args.get("month", type=int) or today.month
-
-    if m < 1 or m > 12:
-        return {"error": "month must be 1..12"}, 400
-
-    last_day = calendar.monthrange(y, m)[1]
-    start = date(y, m, 1)
-    end = date(y, m, last_day)
-
-    try:
-        with SessionLocal() as s:
-            rows = (
-                s.query(FixedAssignment, Staff)
-                .join(Staff, FixedAssignment.staff_id == Staff.id)
-                .filter(FixedAssignment.day.between(start, end))
-                .all()
-            )
-            out = [
-                {
-                    "id": fa.id,
-                    "staff_id": fa.staff_id,
-                    "staff_name": st.full_name,
-                    "day": fa.day.isoformat(),
-                    "shift_code": fa.shift_code,
-                    "position": fa.position,
-                }
-                for fa, st in rows
-            ]
-            return jsonify(out)
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.post("/api/fixed")
-def create_fixed_assignment():
-    data = request.get_json(force=True) or {}
-    staff_id = data.get("staff_id")
-    day_str = data.get("day")
-    shift_code = data.get("shift_code")
-    position = data.get("position")
-
-    if staff_id is None or day_str is None or shift_code is None:
-        return {"error": "staff_id, day, shift_code required"}, 400
-
-    try:
-        staff_id = int(staff_id)
-    except Exception:
-        return {"error": "staff_id must be int"}, 400
-
-    try:
-        d = date.fromisoformat(day_str)
-    except Exception:
-        return {"error": "day must be YYYY-MM-DD"}, 400
-
-    try:
-        with SessionLocal() as s:
-            staff = s.get(Staff, staff_id)
-            if not staff:
-                return {"error": "Staff not found"}, 404
-
-            fa = FixedAssignment(
-                staff_id=staff_id,
-                day=d,
-                shift_code=shift_code,
-                position=position,
-            )
-            s.add(fa)
-            s.commit()
-            item = {
-                "id": fa.id,
-                "staff_id": fa.staff_id,
-                "staff_name": staff.full_name,
-                "day": fa.day.isoformat(),
-                "shift_code": fa.shift_code,
-                "position": fa.position,
-            }
-            return {"ok": True, "item": item}, 201
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.put("/api/fixed/<int:fix_id>")
-def update_fixed_assignment(fix_id: int):
-    data = request.get_json(force=True) or {}
-    staff_id = data.get("staff_id")
-    day_str = data.get("day")
-    shift_code = data.get("shift_code")
-    position = data.get("position")
-
-    try:
-        with SessionLocal() as s:
-            fa = s.get(FixedAssignment, fix_id)
-            if not fa:
-                return {"error": "Not found"}, 404
-
-            if staff_id is not None:
-                try:
-                    staff_id = int(staff_id)
-                except Exception:
-                    return {"error": "staff_id must be int"}, 400
-                staff = s.get(Staff, staff_id)
-                if not staff:
-                    return {"error": "Staff not found"}, 404
-                fa.staff_id = staff_id
-            else:
-                staff = fa.staff
-
-            if day_str is not None:
-                try:
-                    fa.day = date.fromisoformat(day_str)
-                except Exception:
-                    return {"error": "day must be YYYY-MM-DD"}, 400
-
-            if shift_code is not None:
-                fa.shift_code = shift_code
-            if position is not None:
-                fa.position = position
-
-            s.commit()
-            item = {
-                "id": fa.id,
-                "staff_id": fa.staff_id,
-                "staff_name": staff.full_name if staff else None,
-                "day": fa.day.isoformat(),
-                "shift_code": fa.shift_code,
-                "position": fa.position,
-            }
-            return {"ok": True, "item": item}
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.delete("/api/fixed/<int:fix_id>")
-def delete_fixed_assignment(fix_id: int):
-    try:
-        with SessionLocal() as s:
-            fa = s.get(FixedAssignment, fix_id)
-            if not fa:
-                return {"error": "Not found"}, 404
-            s.delete(fa)
-            s.commit()
-            return {"ok": True}
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.get("/api/off")
-def list_off_days():
-    """List off days for a given month (defaults to current)."""
-    today = date.today()
-    y = request.args.get("year", type=int) or today.year
-    m = request.args.get("month", type=int) or today.month
-
-    if m < 1 or m > 12:
-        return {"error": "month must be 1..12"}, 400
-
-    last_day = calendar.monthrange(y, m)[1]
-    start = date(y, m, 1)
-    end = date(y, m, last_day)
-    print(f"[API] off GET {y}-{m}", flush=True)
-
-    try:
-        with SessionLocal() as s:
-            rows = (
-                s.query(OffDay, Staff)
-                .join(Staff, OffDay.staff_id == Staff.id)
-                .filter(OffDay.day.between(start, end))
-                .all()
-            )
-            out = [
-                {
-                    "id": od.id,
-                    "staff_id": od.staff_id,
-                    "staff_name": st.full_name,
-                    "day": od.day.isoformat(),
-                    "reason": od.reason,
-                }
-                for od, st in rows
-            ]
-            return jsonify(out)
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.post("/api/off")
-def create_off_day():
-    data = request.get_json(force=True) or {}
-    staff_id = data.get("staff_id")
-    day_str = data.get("day")
-    reason = data.get("reason")
-
-    if staff_id is None or day_str is None:
-        return {"error": "staff_id and day required"}, 400
-
-    try:
-        staff_id = int(staff_id)
-    except Exception:
-        return {"error": "staff_id must be int"}, 400
-
-    try:
-        d = date.fromisoformat(day_str)
-    except Exception:
-        return {"error": "day must be YYYY-MM-DD"}, 400
-
-    print(f"[API] off POST {staff_id} {day_str}", flush=True)
-
-    try:
-        with SessionLocal() as s:
-            staff = s.get(Staff, staff_id)
-            if not staff:
-                return {"error": "Staff not found"}, 404
-
-            existing = s.query(OffDay).filter(OffDay.staff_id == staff_id, OffDay.day == d).first()
-            if existing:
-                return {"id": existing.id, "ok": True}
-
-            off = OffDay(staff_id=staff_id, day=d, reason=reason)
-            s.add(off)
-            s.commit()
-            return {"id": off.id, "ok": True}, 201
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.delete("/api/off/<int:off_id>")
-def delete_off_day(off_id: int):
-    print(f"[API] off DELETE {off_id}", flush=True)
-    try:
-        with SessionLocal() as s:
-            r = s.get(OffDay, off_id)
-            if not r:
-                return {"error": "Not found"}, 404
-            s.delete(r)
-            s.commit()
-            return {"ok": True}
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-# ---------- Holidays ----------
-@app.get("/api/holidays")
-def list_holidays():
-    """List holidays for a given year, optionally filtering by month."""
-
-    today = date.today()
-    year = request.args.get("year", type=int) or today.year
-    month = request.args.get("month", type=int)
-
-    if month is not None and (month < 1 or month > 12):
-        return {"error": "month must be 1..12"}, 400
-
-    if month is None:
-        start = date(year, 1, 1)
-        end = date(year, 12, 31)
-    else:
-        last_day = calendar.monthrange(year, month)[1]
-        start = date(year, month, 1)
-        end = date(year, month, last_day)
-
-    try:
-        with SessionLocal() as s:
-            rows = (
-                s.query(Holiday)
-                .filter(Holiday.day.between(start, end))
-                .order_by(Holiday.day.asc())
-                .all()
-            )
-            return jsonify([_serialize_holiday(row) for row in rows])
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.post("/api/holidays")
-def create_holiday():
-    data = request.get_json(force=True) or {}
-    day_str = data.get("day")
-    name = data.get("name")
-    kind = data.get("kind") or None
-    official = bool(data.get("official", False))
-    source = data.get("source") or None
-
-    if not day_str:
-        return {"error": "day required"}, 400
-
-    try:
-        holiday_day = date.fromisoformat(day_str)
-    except Exception:
-        return {"error": "day must be YYYY-MM-DD"}, 400
-
-    try:
-        with SessionLocal() as s:
-            existing = s.query(Holiday).filter(Holiday.day == holiday_day).first()
-            if existing:
-                return {"ok": True, "item": _serialize_holiday(existing)}
-
-            holiday = Holiday(
-                day=holiday_day,
-                name=name,
-                kind=kind,
-                official=official,
-                source=source,
-            )
-            s.add(holiday)
-            s.commit()
-            s.refresh(holiday)
-            return {"ok": True, "item": _serialize_holiday(holiday)}, 201
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.put("/api/holidays/<int:holiday_id>")
-def update_holiday(holiday_id: int):
-    data = request.get_json(force=True) or {}
-
-    try:
-        with SessionLocal() as s:
-            row = s.get(Holiday, holiday_id)
-            if not row:
-                return {"error": "Not found"}, 404
-
-            if "day" in data:
-                day_str = data.get("day")
-                try:
-                    new_day = date.fromisoformat(day_str)
-                except Exception:
-                    return {"error": "day must be YYYY-MM-DD"}, 400
-
-                conflict = (
-                    s.query(Holiday)
-                    .filter(Holiday.day == new_day, Holiday.id != holiday_id)
-                    .first()
-                )
-                if conflict:
-                    return {"error": "day already exists"}, 409
-                row.day = new_day
-
-            if "name" in data:
-                row.name = data.get("name")
-            if "kind" in data:
-                row.kind = data.get("kind") or None
-            if "official" in data:
-                row.official = bool(data.get("official"))
-            if "source" in data:
-                row.source = data.get("source") or None
-
-            s.commit()
-            s.refresh(row)
-            return {"ok": True, "item": _serialize_holiday(row)}
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.delete("/api/holidays/<int:holiday_id>")
-def delete_holiday(holiday_id: int):
-    try:
-        with SessionLocal() as s:
-            row = s.get(Holiday, holiday_id)
-            if not row:
-                return {"error": "Not found"}, 404
-            s.delete(row)
-            s.commit()
-            return {"ok": True}
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-@app.post("/api/holidays/import")
-def import_holidays():
-    """Import holidays from external providers."""
-
-    year = request.args.get("year", type=int)
-    source = (request.args.get("source") or "").lower()
-
-    if not year:
-        return {"error": "year required"}, 400
-    if source != "nager":
-        return {"error": "unsupported source"}, 400
-
-    try:
-        incoming = _fetch_nager_holidays(year)
-    except HolidayImportError as exc:
-        return {"error": f"Import failed: {exc}"}, 502
-
-    days = [item["day"] for item in incoming]
-
-    try:
-        with SessionLocal() as s:
-            existing_rows = (
-                s.query(Holiday)
-                .filter(Holiday.day.in_(days))
-                .all()
-                if days
-                else []
-            )
-            existing_by_day = {row.day: row for row in existing_rows}
-
-            inserted = 0
-            updated = 0
-
-            for item in incoming:
-                existing = existing_by_day.get(item["day"])
-                if existing:
-                    existing.name = item["name"]
-                    existing.kind = item["kind"]
-                    existing.official = item["official"]
-                    existing.source = item["source"]
-                    updated += 1
-                else:
-                    holiday = Holiday(
-                        day=item["day"],
-                        name=item["name"],
-                        kind=item["kind"],
-                        official=item["official"],
-                        source=item["source"],
-                    )
-                    s.add(holiday)
-                    inserted += 1
-
-            s.commit()
-
-            year_rows = (
-                s.query(Holiday)
-                .filter(
-                    Holiday.day.between(
-                        date(year, 1, 1),
-                        date(year, 12, 31),
-                    )
-                )
-                .order_by(Holiday.day.asc())
-                .all()
-            )
-
-            return {
-                "ok": True,
-                "inserted": inserted,
-                "updated": updated,
-                "items": [_serialize_holiday(row) for row in year_rows],
-            }
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}),
-            500,
-        )
-
-
-# ---------- Shift plan defaults ----------
-@app.get("/api/shift-defaults")
-def get_shift_defaults():
-    try:
-        year, month = _parse_year_month(
-            request.args.get("year"), request.args.get("month")
-        )
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-    except Exception as exc:  # pragma: no cover - defensive guard
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": f"Internal error: {exc.__class__.__name__}: {exc}",
-                }
-            ),
-            500,
-        )
-
-    try:
-        with SessionLocal() as session:
-            defaults = session.execute(
-                select(ShiftPlanDefaults).where(
-                    ShiftPlanDefaults.year == year,
-                    ShiftPlanDefaults.month == month,
-                )
-            ).scalar_one_or_none()
-            return jsonify(_serialize_shift_defaults(defaults, year, month))
-    except Exception as exc:  # pragma: no cover - defensive guard
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": f"Internal error: {exc.__class__.__name__}: {exc}",
-                }
-            ),
-            500,
-        )
-
-
-@app.put("/api/shift-defaults")
-def upsert_shift_defaults():
-    data = request.get_json(silent=True) or {}
-    try:
-        year, month = _parse_year_month(data.get("year"), data.get("month"))
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-    except Exception as exc:  # pragma: no cover - defensive guard
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": f"Internal error: {exc.__class__.__name__}: {exc}",
-                }
-            ),
-            500,
-        )
-
-    fields = ["day_shifts", "night_shifts", "leader_shifts", "pgd_shifts"]
-    missing = [field for field in fields if field not in data]
-    if missing:
-        return {"error": f"Missing fields: {', '.join(missing)}"}, 400
-
-    validated: dict[str, int] = {}
-    for field in fields:
-        value = data.get(field)
-        if not isinstance(value, int):
-            return {"error": f"{field} must be an integer"}, 400
-        if value < 0:
-            return {"error": f"{field} must be >= 0"}, 400
-        validated[field] = value
-
-    try:
-        with SessionLocal() as session:
-            defaults = session.execute(
-                select(ShiftPlanDefaults).where(
-                    ShiftPlanDefaults.year == year,
-                    ShiftPlanDefaults.month == month,
-                )
-            ).scalar_one_or_none()
-
-            if defaults is None:
-                defaults = ShiftPlanDefaults(
-                    year=year,
-                    month=month,
-                    day_shifts=validated["day_shifts"],
-                    night_shifts=validated["night_shifts"],
-                    leader_shifts=validated["leader_shifts"],
-                    pgd_shifts=validated["pgd_shifts"],
-                )
-                session.add(defaults)
-            else:
-                for field, value in validated.items():
-                    setattr(defaults, field, value)
-
-            session.commit()
-            session.refresh(defaults)
-            return jsonify(_serialize_shift_defaults(defaults, year, month))
-    except Exception as exc:  # pragma: no cover - defensive guard
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": f"Internal error: {exc.__class__.__name__}: {exc}",
-                }
-            ),
-            500,
-        )
-
-
-# ---------- Month config ----------
-@app.get("/api/month-config")
-def get_month_config():
-    try:
-        year, month = _parse_year_month(
-            request.args.get("year"), request.args.get("month")
-        )
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-    except Exception as exc:  # pragma: no cover - defensive guard
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": f"Internal error: {exc.__class__.__name__}: {exc}",
-                }
-            ),
-            500,
-        )
-
-    try:
-        with SessionLocal() as session:
-            payload = _build_month_config_payload(session, year, month)
-            return jsonify(payload)
-    except Exception as exc:  # pragma: no cover - defensive guard
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": f"Internal error: {exc.__class__.__name__}: {exc}",
-                }
-            ),
-            500,
-        )
-
-
-@app.put("/api/month-config")
-def upsert_month_config():
-    data = request.get_json(silent=True) or {}
-    try:
-        year, month = _parse_year_month(data.get("year"), data.get("month"))
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-    except Exception as exc:  # pragma: no cover - defensive guard
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": f"Internal error: {exc.__class__.__name__}: {exc}",
-                }
-            ),
-            500,
-        )
-
-    try:
-        weekend_policy = None
-        working_days_override = None
-
-        def _existing_dates(values: list[str] | None) -> set[date]:
-            dates: set[date] = set()
-            for raw in values or []:
-                if not isinstance(raw, str):
-                    continue
-                try:
-                    dt = date.fromisoformat(raw)
-                except ValueError:
-                    continue
-                if dt.year == year and dt.month == month:
-                    dates.add(dt)
-            return dates
-
-        with SessionLocal() as session:
-            config = session.execute(
-                select(MonthConfig).where(
-                    MonthConfig.year == year, MonthConfig.month == month
-                )
-            ).scalar_one_or_none()
-
-            if config is None:
-                config = MonthConfig(year=year, month=month)
-                session.add(config)
-
-            if "weekend_policy" in data:
-                policy_raw = data.get("weekend_policy")
-                if policy_raw is None:
-                    weekend_policy = WeekendPolicy.SAT_OFF
-                else:
-                    try:
-                        weekend_policy = WeekendPolicy(policy_raw)
-                    except ValueError as exc:  # pragma: no cover - enum guard
-                        raise ValueError("weekend_policy invalid") from exc
-            else:
-                weekend_policy = config.weekend_policy or WeekendPolicy.SAT_OFF
-
-            override_sentinel = object()
-            override_raw = data.get("working_days_override", override_sentinel)
-            if override_raw is override_sentinel:
-                working_days_override = config.working_days_override
-            elif override_raw is None:
-                working_days_override = None
-            else:
-                try:
-                    working_days_override = int(override_raw)
-                except (TypeError, ValueError) as exc:  # pragma: no cover - guard
-                    raise ValueError("working_days_override must be an integer") from exc
-                if working_days_override < 0:
-                    raise ValueError("working_days_override must be >= 0")
-
-            existing_workdays = _existing_dates(config.extra_workdays)
-            existing_offdays = _existing_dates(config.extra_offdays)
-
-            workdays = _coerce_extra_dates(
-                data.get("extra_workdays"), year, month, "extra_workdays"
-            )
-            offdays = _coerce_extra_dates(
-                data.get("extra_offdays"), year, month, "extra_offdays"
-            )
-
-            if workdays is None:
-                workdays = set(existing_workdays)
-            else:
-                workdays = set(workdays)
-            if offdays is None:
-                offdays = set(existing_offdays)
-            else:
-                offdays = set(offdays)
-
-            workdays -= offdays
-
-            config.year = year
-            config.month = month
-            config.weekend_policy = weekend_policy
-            config.extra_workdays = sorted(dt.isoformat() for dt in workdays)
-            config.extra_offdays = sorted(dt.isoformat() for dt in offdays)
-            config.working_days_override = working_days_override
-
-            session.commit()
-
-            payload = _build_month_config_payload(session, year, month)
-            return jsonify(payload)
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-
-
-# Backwards‑compat aliases for old /api/offdays routes
-@app.get("/api/offdays")
-def list_off_days_alias():
-    return list_off_days()
-
-
-@app.post("/api/offdays")
-def create_off_day_alias():
-    return create_off_day()
-
-
-@app.delete("/api/offdays/<int:off_id>")
-def delete_off_day_alias(off_id: int):
-    return delete_off_day(off_id)
-
-
-@app.get("/api/assignments")
-def list_assignments():
-    """Trả danh sách phân ca theo tháng (mặc định tháng hiện tại nếu thiếu params)."""
-    today = date.today()
-    y = request.args.get("year", type=int) or today.year
-    m = request.args.get("month", type=int) or today.month
-
-    if m < 1 or m > 12:
-        return {"error": "month must be 1..12"}, 400
-
-    last_day = calendar.monthrange(y, m)[1]
-    start = date(y, m, 1)
-    end = date(y, m, last_day)
-
-    with SessionLocal() as s:
-        rows = s.query(Assignment).filter(Assignment.day.between(start, end)).all()
-        return jsonify(
-            [
-                {
-                    "day": r.day.isoformat(),
-                    "shift_code": r.shift_code,
-                    "staff_id": r.staff_id,
-                    "position": r.position,  # PGD | TD | None
-                }
-                for r in rows
-            ]
-        )
-
-
-@app.get("/api/export_audit")
-def export_audit():
-    """Xuất danh sách phân ca kèm JSON metadata ra CSV để audit.
-
-    Sử dụng ``csv.writer`` để đảm bảo trích dẫn đúng các ký tự đặc biệt
-    trong cột JSON. Dữ liệu được stream từng hàng thông qua ``io.StringIO``
-    nhằm tránh giữ toàn bộ nội dung trong bộ nhớ.
-    """
-    today = date.today()
-    y = request.args.get("year", type=int) or today.year
-    m = request.args.get("month", type=int) or today.month
-
-    if m < 1 or m > 12:
-        return {"error": "month must be 1..12"}, 400
-
-    last_day = calendar.monthrange(y, m)[1]
-    start = date(y, m, 1)
-    end = date(y, m, last_day)
-
-    with SessionLocal() as s:
-        rows = s.query(Assignment).filter(Assignment.day.between(start, end)).all()
-
-    def generate():
-        buf = io.StringIO()
-        writer = csv.writer(buf)
-        writer.writerow(["day", "shift_code", "staff_id", "meta"])
-        yield buf.getvalue()
-        buf.seek(0)
-        buf.truncate(0)
-        for r in rows:
-            # JSON chứa nhiều key, bao gồm chuỗi có dấu phẩy để kiểm tra quoting
-            meta = {
-                "info": f"{r.shift_code},{r.staff_id}",
-                "extra": "value",
-            }
-            writer.writerow(
-                [
-                    r.day.isoformat(),
-                    r.shift_code,
-                    r.staff_id,
-                    json.dumps(meta, ensure_ascii=False),
-                ]
-            )
-            yield buf.getvalue()
-            buf.seek(0)
-            buf.truncate(0)
-
-    headers = {"Content-Disposition": f"attachment; filename=audit-{y:04d}-{m:02d}.csv"}
-    return Response(
-        stream_with_context(generate()),
-        content_type="text/csv; charset=utf-8",
-        headers=headers,
-    )
-
-
-@app.get("/api/export/month.csv")
-def export_month_csv_endpoint():
-    """Xuất lịch phân ca dạng CSV cho một tháng."""
-    today = date.today()
-    y = request.args.get("year", type=int) or today.year
-    m = request.args.get("month", type=int) or today.month
-    if m < 1 or m > 12:
-        return {"error": "month must be 1..12"}, 400
-    return _export_month_csv(y, m)
-
-
-@app.get("/api/metrics/staff-workload")
-def metrics_staff_workload():
-    try:
-        year, month = _parse_year_month(
-            request.args.get("year"), request.args.get("month")
-        )
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-
-    data, totals = load_staff_workload(year, month)
-    return jsonify(
-        {
-            "by_staff": [
-                {
-                    "staff_id": row.staff_id,
-                    "name": row.name,
-                    "hours": row.hours,
-                    "night_hours": row.night_hours,
-                }
-                for row in data
-            ],
-            "totals": totals,
-        }
-    )
-
-
-@app.get("/api/metrics/department-compare")
-def metrics_department_compare():
-    try:
-        year, month = _parse_year_month(
-            request.args.get("year"), request.args.get("month")
-        )
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-
-    rows = load_department_comparison(year, month)
-    return jsonify(
-        {
-            "by_department": [
-                {
-                    "dept": row.dept,
-                    "staff_count": row.staff_count,
-                    "hours": row.hours,
-                    "overtime_hours": row.overtime_hours,
-                }
-                for row in rows
-            ]
-        }
-    )
-
-
-@app.get("/api/metrics/attendance")
-def metrics_attendance():
-    start_raw = request.args.get("from")
-    end_raw = request.args.get("to")
-    if not start_raw or not end_raw:
-        return {"error": "from and to required"}, 400
-
-    try:
-        start = date.fromisoformat(start_raw)
-        end = date.fromisoformat(end_raw)
-    except ValueError:
-        return {"error": "from/to must be YYYY-MM-DD"}, 400
-
-    if end < start:
-        return {"error": "to must be on or after from"}, 400
-
-    return jsonify({"rate": 0, "absences": [], "late": []})
-
-
-@app.get("/api/metrics/cost")
-def metrics_cost():
-    try:
-        year, month = _parse_year_month(
-            request.args.get("year"), request.args.get("month")
-        )
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-
-    _, totals = load_staff_workload(year, month)
-    cost = totals["hours"] * _get_cost_per_hour()
-    return jsonify({"labor_cost": cost})
-
-
-@app.get("/api/reports/staff-workload.csv")
-def export_staff_workload_csv():
-    try:
-        year, month = _parse_year_month(
-            request.args.get("year"), request.args.get("month")
-        )
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-
-    data, totals = load_staff_workload(year, month)
-    rows = [
-        [row.staff_id, row.name, f"{row.hours:.2f}", f"{row.night_hours:.2f}"]
-        for row in data
-    ]
-    if data:
-        rows.append(["TOTAL", "", f"{totals['hours']:.2f}", f"{totals['night_hours']:.2f}"])
-    return _build_csv_response(
-        f"staff-workload-{year:04d}-{month:02d}.csv",
-        ["staff_id", "name", "hours", "night_hours"],
-        rows,
-    )
-
-
-@app.get("/api/reports/department-compare.csv")
-def export_department_compare_csv():
-    try:
-        year, month = _parse_year_month(
-            request.args.get("year"), request.args.get("month")
-        )
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-
-    rows = load_department_comparison(year, month)
-    csv_rows = [
-        [
-            row.dept,
-            row.staff_count,
-            f"{row.hours:.2f}",
-            f"{row.overtime_hours:.2f}",
-        ]
-        for row in rows
-    ]
-    return _build_csv_response(
-        f"department-compare-{year:04d}-{month:02d}.csv",
-        ["department", "staff_count", "hours", "overtime_hours"],
-        csv_rows,
-    )
-
-
-@app.get("/api/reports/schedule-month.csv")
-def export_schedule_month_csv():
-    try:
-        year, month = _parse_year_month(
-            request.args.get("year"), request.args.get("month")
-        )
-    except ValueError as exc:
-        return {"error": str(exc)}, 400
-
-    response = _export_month_csv(year, month)
-    response.headers["Content-Disposition"] = (
-        f"attachment; filename=schedule-month-{year:04d}-{month:02d}.csv"
-    )
-    return response
-
-
-@app.get("/api/rules/expected")
-def rule_expected():
-    today = date.today()
-    y = request.args.get("year", type=int) or today.year
-    m = request.args.get("month", type=int) or today.month
-    prof = get_profile()
-
-    first = ymd(y, m, 1)
-    last = ymd(y, m, month_last_day(y, m))
-
-    out = {}
-    d = first
-    # nạp holiday 1 lần
-    holidays = load_holidays(SessionLocal())
-    while d <= last:
-        kind = day_kind(d, holidays)
-
-        day = prof.expected_day_counts(
-            kind
-        )  # {"TD": {"K":..,"CA1":..,"CA2":..}, "PGD": {"K":..,"CA2":..}}
-        night = prof.expected_night_counts(
-            kind
-        )  # {"TD": {"Đ": x}, "PGD": {"Đ": y}}  (từ NightDetail.to_engine_dict)
-
-        td = day.get("TD", {})
-        pgd = day.get("PGD", {})
-        n_td = night.get("TD", {}) or {}
-        n_pgd = night.get("PGD", {}) or {}
-
-        out[d.day] = {
-            "expectedTD": {
-                "K": int(td.get("K", 0)),
-                "CA1": int(td.get("CA1", 0)),
-                "CA2": int(td.get("CA2", 0)),
-                "D": int(n_td.get("Đ", 0)),  # gộp Đ (đêm @ TD)
-            },
-            "expectedPGD": {
-                "K": int(pgd.get("K", 0)),
-                "CA2": int(pgd.get("CA2", 0)),
-                "D": int(n_pgd.get("Đ", 0)),  # gộp Đ (đêm @ PGD)
-            },
-        }
-        d = d + timedelta(days=1)
-
-    return jsonify({"ok": True, "perDayExpected": out})
-
-
-@app.get("/api/schedule/validate")
-def schedule_validate():
-    today = date.today()
-    y = request.args.get("year", type=int) or today.year
-    m = request.args.get("month", type=int) or today.month
-    if m < 1 or m > 12:
-        return jsonify({"ok": False, "error": "month must be 1..12"}), 400
-    body = validate_month(y, m)
-    return jsonify(body)
-
-
-@app.post("/api/schedule/generate")
-def gen():
-    payload = request.get_json(silent=True) or {}
-    from datetime import date as _date
-
-    today = _date.today()
-    year = int(payload.get("year", today.year))
-    month = int(payload.get("month", today.month))
-    if month < 1 or month > 12:
-        return jsonify({"ok": False, "error": "month must be 1..12"}), 400
-
-    shuffle = bool(payload.get("shuffle", False))
-    seed = payload.get("seed")
-    try:
-        seed = int(seed) if seed is not None else None
-    except Exception:
-        seed = None
-    save_flag = bool(payload.get("save", False))
-    fill_hc = bool(payload.get("fill_hc", False))
-
-    try:
-        with SessionLocal() as session:
-            month_config = session.execute(
-                select(MonthConfig).where(
-                    MonthConfig.year == year, MonthConfig.month == month
-                )
-            ).scalar_one_or_none()
-            shift_defaults = session.execute(
-                select(ShiftPlanDefaults).where(
-                    ShiftPlanDefaults.year == year,
-                    ShiftPlanDefaults.month == month,
-                )
-            ).scalar_one_or_none()
-
-            missing: dict[str, str] = {}
-            if month_config is None:
-                missing["month_config"] = (
-                    f"No month config found for {year}-{month:02d}"
-                )
-            if shift_defaults is None:
-                missing["shift_plan_defaults"] = (
-                    f"No shift plan defaults found for {year}-{month:02d}"
-                )
-            if missing:
-                return (
-                    jsonify(
-                        {
-                            "ok": False,
-                            "error": "Missing configuration",
-                            "missing": missing,
-                        }
-                    ),
-                    400,
-                )
-
-            config_payload = _build_month_config_payload(session, year, month)
-            effective_working_days = float(
-                config_payload.get("effective_working_days", 0.0)
-            )
-            shift_plan_defaults = {
-                "day_shifts": shift_defaults.day_shifts,
-                "night_shifts": shift_defaults.night_shifts,
-                "leader_shifts": shift_defaults.leader_shifts,
-                "pgd_shifts": shift_defaults.pgd_shifts,
-            }
-    except Exception as exc:  # pragma: no cover - defensive guard
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": f"Internal error: {exc.__class__.__name__}: {exc}",
-                }
-            ),
-            500,
-        )
-
-    print("[API] generate begin", year, month, shuffle, seed, save_flag, fill_hc, flush=True)
-
-    try:
-        res = generate_schedule(
-            year,
-            month,
-            shuffle=shuffle,
-            seed=seed,
-            save=save_flag,
-            fill_hc=fill_hc,
-            effective_working_days=effective_working_days,
-            shift_plan_defaults=shift_plan_defaults,
-        )
-        print("[API] generate end", type(res), flush=True)
-        # engine may return a dict or (dict, status)
-        if isinstance(res, tuple) and len(res) == 2 and isinstance(res[0], dict):
-            body, status = res
-            return jsonify(body), int(status)
-        elif isinstance(res, dict):
-            return jsonify(res)
-        else:
-            return jsonify({"ok": False, "error": "Internal error: bad return type"}), 500
-    except Exception as e:
-        # never bubble raw HTML traceback to the frontend
-        return jsonify({"ok": False, "error": f"Internal error: {e.__class__.__name__}: {e}"}), 500
-
-
-# ====== Reset DB ======
-@app.post("/api/admin/reset")
-def admin_reset():
-    """
-    mode=soft: xoá Assignment (giữ Staff/Off/Fix/Holiday)
-    mode=hard: xoá file sqlite và tạo lại schema
-    """
-    mode = (request.args.get("mode") or "soft").lower()
-    if mode not in ("soft", "hard"):
-        return {"error": "mode must be soft|hard"}, 400
-
-    if mode == "hard":
-        try:
-            if DB_FILE and DB_FILE.exists():
-                DB_FILE.unlink()
-            init_db()  # tạo lại schema rỗng
-            return {"ok": True, "mode": "hard"}
-        except Exception as e:
-            return {"error": str(e)}, 500
-
-    # soft reset
-    with SessionLocal() as s:
-        try:
-            s.execute(text("DELETE FROM assignment"))
-            s.commit()
-            return {"ok": True, "mode": "soft"}
-        except Exception as e:
-            s.rollback()
-            return {"error": str(e)}, 500
-
-
-@app.get("/api/schedule/estimate")
-def api_estimate():
-    today = date.today()
-    y = request.args.get("year", type=int) or today.year
-    m = request.args.get("month", type=int) or today.month
-    return jsonify(estimate_month(y, m))
-
-
-@app.post("/api/chatbot/upload")
-def api_chatbot_upload():
-    """Proxy endpoint to forward file uploads to the webhook"""
-    import requests
-
-    if 'file' not in request.files:
-        return jsonify({"error": "No file provided"}), 400
-
-    file = request.files['file']
-    if file.filename == '':
-        return jsonify({"error": "No file selected"}), 400
-
-    try:
-        # Forward the file to the webhook
-        files = {'file': (file.filename, file.stream, file.content_type)}
-
-        # Include any additional form data
-        data = {}
-        for key in request.form:
-            data[key] = request.form[key]
-
-        webhook_url = os.getenv("WEBHOOK_UPLOAD_URL", "https://iconic-host.lapage.vn/webhook/upload-file")
-        response = requests.post(webhook_url, files=files, data=data, timeout=30)
-
-        return jsonify({
-            "status": response.status_code,
-            "response": response.text
-        }), response.status_code
-
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@app.post("/api/chatbot/chunking")
-def api_chatbot_chunking():
-    """Proxy endpoint to forward chunking request to the webhook"""
-    import requests
-
-    try:
-        data = request.get_json()
-        if not data:
-            return jsonify({"error": "No data provided"}), 400
-
-        webhook_url = os.getenv("WEBHOOK_CHUNKING_URL", "https://iconic-host.lapage.vn/webhook/chunking")
-        response = requests.post(webhook_url, json=data, timeout=30)
-
-        return jsonify({
-            "status": response.status_code,
-            "response": response.json() if response.headers.get('content-type') == 'application/json' else response.text
-        }), response.status_code
-
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
+# Register middleware and blueprints
+register_api_key_middleware(app, api_key=API_KEY)
+register_routes(app)
 
 
 if __name__ == "__main__":
     app.run(host=APP_HOST, port=APP_PORT, debug=APP_DEBUG)
+
